Refactor #4 - Move detail_xml to detail.py
