+++ judgments/urls.py
@@ -3,7 +3,7 @@ from django.urls import path, re_path, register_converter
 from . import converters, feeds, views
 from .views.advanced_search import advanced_search
 from .views.browse import browse
-from .views.detail import detail
+from .views.detail import detail, detail_xml
 
 register_converter(converters.YearConverter, "yyyy")
 register_converter(converters.DateConverter, "date")
@@ -46,9 +46,7 @@ urlpatterns = [
         views.PdfDetailView.as_view(),
         name="weasy_pdf",
     ),
-    re_path(
-        r"(?P<judgment_uri>.*/\d{4}/\d+)/data.xml", views.detail_xml, name="detail_xml"
-    ),
+    re_path(r"(?P<judgment_uri>.*/\d{4}/\d+)/data.xml", detail_xml, name="detail_xml"),
     re_path(r"(?P<judgment_uri>.*/\d{4}/\d+)/data.html", detail, name="detail"),
     re_path(r"(?P<judgment_uri>.*/\d{4}/\d+)", detail, name="detail"),
     path("judgments/results", views.results, name="results"),
